+++ gral-core/src/main/java/de/erichseifert/gral/plots/colors/Grayscale.java
@@ -46,7 +46,7 @@ public class Grayscale extends ScaledContinuousColorMapper {
 		if (!MathUtils.isCalculatable(v)) {
 			return null;
 		}
-		double lightness = 100.0*v.doubleValue();
+		double lightness = 100.0*v;
 		double[] rgb = GraphicsUtils.luv2rgb(new double[] {lightness, 0.0, 0.0}, null);
 		return new Color(
 			(float) MathUtils.limit(rgb[0], 0.0, 1.0),
